dev alaa

+++ modules/users/controller/user.controller.js
@@ -2,6 +2,7 @@ const { validationResult } = require('express-validator');
 const UserInput = require('../inputs/user.input');
 const { UserRegister, getAllUsers, GetUser, Updateuser, DeleteUserService } = require('../services/user.service');
 const UpdateUserInput = require('../inputs/update.user.input');
+const { ADMIN, USER } = require('../../Roles/roles')
 //user register
 const UserRegisterController = async (req, res) => {
     const errors = validationResult(req);
@@ -11,11 +12,12 @@ const UserRegisterController = async (req, res) => {
         })
     }
     const user = new UserInput()
-    const { name, email, password, phone } = req.body;
+    const { name, email, password, phone, role } = req.body;
     user.name = name;
     user.email = email;
     user.password = password;
     user.phone = phone;
+    user.role = role;
     let RegisterdUser = await UserRegister(user);
     res.status(201).json({
         result: RegisterdUser
@@ -23,10 +25,17 @@ const UserRegisterController = async (req, res) => {
 }
 //get all users controller 
 const GetAllUsersController = async (req, res) => {
-    const users = await getAllUsers();
-    res.status(200).json({
-        users: users
-    })
+    const { role } = req.params;
+    if (role === ADMIN) {
+        const users = await getAllUsers();
+        res.status(200).json({
+            users: users
+        })
+    } else {
+        res.json({
+            error: "you don't have any permissions"
+        })
+    }
 }
 //get user by id 
 const getUserById = async (req, res) => {
@@ -56,11 +65,17 @@ const UpdateUserById = async (req, res) => {
 }
 //delete user by id 
 const DeleteUserById = async (req, res) => {
-    const { user_id } = req.params;
-    const deletedUser = await DeleteUserService(user_id);
-    res.status(200).json({
-        success: true
-    })
+    const { user_id, role } = req.params;
+    if (role === ADMIN) {
+        const deletedUser = await DeleteUserService(user_id);
+        res.status(200).json({
+            success: true
+        })
+    } else {
+        res.json({
+            message: "Tou don't have permissions"
+        })
+    }
 }
 module.exports = {
     UserRegisterController,
